+++ core/wma/inc/wma_tgt_cfg.h
@@ -157,7 +157,7 @@ struct board_info {
  * struct wma_tgt_cfg - target config
  * @target_fw_version: target fw version
  * @target_fw_vers_ext: target fw extended sub version
- * @band_cap: band capability
+ * @band_cap: band capability bitmap
  * @reg_domain: reg domain
  * @eeprom_rd_ext: eeprom rd ext
  * @hw_macaddr: hw mcast addr
@@ -182,7 +182,7 @@ struct board_info {
 struct wma_tgt_cfg {
 	uint32_t target_fw_version;
 	uint32_t target_fw_vers_ext;
-	uint8_t band_cap;
+	uint32_t band_cap;
 	uint32_t reg_domain;
 	uint32_t eeprom_rd_ext;
 	struct qdf_mac_addr hw_macaddr;
